Removed cloning/refcounting from Property s
Turns out that it's useful if a Property has a handle to the SlideElement that
it is being used for. While I could keep the refcounting stuff for memory
purposes, it's a lot of overhead for only a small bit of memory gain. You would
probably need to have a PropertyModel that could be shared amongst multiple
Property objects, which could each hold a handle to the individual SlideElement
they were registered with. Seems like overkill to me unless a Property is really
memory heavy